single manager to rule them all
